requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
Django==4.1.6
django-cors-headers==3.13.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
graphene==3.2.2
graphene-django==3.0.1
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
phonenumbers==8.13.7
Pillow==9.4.0
promise==2.3
psycopg2-binary==2.9.6
PyJWT==2.6.0
pyotp==2.8.0
pytz==2022.7.1
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
typing==3.7.4.3
